Refactor: Merge all Python modules into a single file and add test cases
- Combined `layout_parser.py` and `layout_visualizer.py` into a single file `GenGIF.py`.
- Removed unused files: `requirements.txt` and `utils.py`.
- Simplified the structure for easier usage and maintenance.
- Added `testcase0` and `testcase0_c` along with their `.lg`, `.opt`, and `.post.lg` files.
- Included generated GIFs (`testcase0.gif` and `testcase0_c.gif`) as outputs for demonstration.
