--- GenGIF/requirements.txt
@@ -1,3 +1,0 @@
-matplotlib
-Pillow
-argparse
