+++ GenGIF/testcase0.lg
@@ -1,0 +1,20 @@
+Alpha 100
+Beta 200
+DieSize 0 0 100 100
+FF_1_0 5 0 5 10 NOTFIX
+FF_1_1 10 0 5 10 NOTFIX
+FF_1_2 10 20 5 10 NOTFIX
+FF_1_3 40 10 10 20 NOTFIX
+FF_1_4 50 30 10 10 NOTFIX
+FF_1_5 60 0 5 10 NOTFIX
+FF_1_6 70 20 5 10 NOTFIX
+C4 10 10 5 10 FIX
+Gate_1 20 30 10 10 FIX
+Gate_2 40 40 20 20 FIX
+PlacementRows 0 0 1 10 100
+PlacementRows 0 10 1 10 100
+PlacementRows 0 20 1 10 100
+PlacementRows 0 30 1 10 100
+PlacementRows 0 40 1 10 100
+PlacementRows 0 50 1 10 100
+PlacementRows 0 60 1 10 100
