+++ GenGIF/testcase0_c.lg
@@ -1,0 +1,23 @@
+Alpha 100
+Beta 200
+DieSize 0 0 50 50
+FF_1_0 5 0 5 10 NOTFIX
+FF_1_1 10 0 5 10 NOTFIX
+FF_1_2 20 0 5 10 NOTFIX
+FF_1_3 30 0 5 10 NOTFIX
+FF_1_4 0 10 5 10 NOTFIX
+FF_1_5 10 10 5 10 NOTFIX
+FF_1_6 20 10 5 10 NOTFIX
+FF_1_7 30 10 5 10 NOTFIX
+FF_1_8 0 20 5 10 NOTFIX
+FF_1_9 10 20 5 10 NOTFIX
+FF_1_10 20 20 5 10 NOTFIX
+C4_1 40 0 5 10 FIX
+C4_2 40 10 5 10 FIX
+Gate_1 0 30 10 10 FIX
+Gate_2 20 30 10 10 FIX
+Gate_3 40 30 10 10 FIX
+PlacementRows 0 0 1 10 50
+PlacementRows 0 10 1 10 50
+PlacementRows 0 20 1 10 50
+PlacementRows 0 30 1 10 50
