+++ GenGIF/testcase0_c_post.lg
@@ -1,0 +1,12 @@
+10 0
+0
+20 0
+0
+10 10
+0
+20 10
+0
+10 20
+0
+20 20
+0
