+++ GenGIF/testcase0_post.lg
@@ -1,0 +1,8 @@
+65 0
+0
+20 20
+0
+60 30
+0
+40 10
+0
