--- GenGIF/utils.py
@@ -1,3 +1,0 @@
-def normalize_coordinates(layout, scale=1):
-    # Normalize coordinates for better visualization
-    pass
